Continuation du rapport + petites modifs de type pour supporter 2^32

+++ fonctions_individu/decoder_indiv.c
@@ -2,8 +2,8 @@
 #include "longueur_indiv.h"
 #include "../fonctions_math/puissance.h"
 
-int decoder_indiv(individu personne) {
-    int somme = 0;
+unsigned long long decoder_indiv(individu personne) {
+    unsigned long long somme = 0;
 
     for (int i = longueur_indiv(personne) - 1; i >= 0; i--) {
         somme += (personne->valeur - '0')*puissance(2, i);
