+++ fonctions_math/puissance.h
@@ -1,7 +1,7 @@
 #ifndef PUISSANCE_H_INCLUDED
 #define PUISSANCE_H_INCLUDED
 
-// Calcule x^n de manière récursive pour n >= 0.
-int puissance(int, int);
+// Calcule x^n de manière itérative pour n >= 0 et x >= 0.
+unsigned long long puissance(int, int);
 
 #endif // PUISSANCE_H_INCLUDED
